+++ R/tab-book.R
@@ -120,6 +120,17 @@ setMethod("dim", "TabBookResult", function (x) {
 setMethod("names", "TabBookResult", function (x) {
     unlist(lapply(x$meta$sheets, function (sheet) sheet$name))
 })
+#' @rdname tabbook-methods
+#' @export
+setMethod("aliases", "TabBookResult", function (x) {
+    unlist(lapply(x, function (mt) aliases(mt[[1]])[1]), use.names=FALSE)
+})
+#' @rdname tabbook-methods
+#' @export
+setMethod("descriptions", "TabBookResult", function (x) {
+    unlist(lapply(x, function (mt) descriptions(mt[[1]])[1]), use.names=FALSE)
+})
+
 setMethod("lapply", "TabBookResult", function (X, FUN, ...) {
     lapply(X$sheets, FUN, ...)
 })
@@ -147,7 +158,10 @@ setMethod("initialize", "MultitableResult", function (.Object, ...) {
         ## that column is multitable var (3 -> 2), row is category of
         ## array (2 -> 1), subvar is "tab" (1 -> 3)
         if (length(dim(cube)) == 3L) {
-            cube@dims <- CubeDims(cube@dims[c(2, 3, 1)])
+            ## TODO: CubeDims should get a [ method
+            ## TODO: CrunchCube should get a dimensions<-
+            cube@dims <- CubeDims(dimensions(cube)[c(2, 3, 1)],
+                references=dimensions(cube)@references[c(2, 3, 1)])
             cube@arrays <- lapply(cube@arrays, aperm, perm=c(2, 3, 1))
         }
         return(cube)
@@ -164,6 +178,21 @@ setMethod("[[", "MultitableResult", function (x, i, ...) {
 })
 setMethod("lapply", "MultitableResult", function (X, FUN, ...) {
     lapply(X$result, FUN, ...)
+})
+#' @rdname tabbook-methods
+#' @export
+setMethod("names", "MultitableResult", function (x) {
+    unlist(lapply(x, function (cube) names(cube)[2]), use.names=FALSE)
+})
+#' @rdname tabbook-methods
+#' @export
+setMethod("aliases", "MultitableResult", function (x) {
+    unlist(lapply(x, function (cube) aliases(cube)[2]), use.names=FALSE)
+})
+#' @rdname tabbook-methods
+#' @export
+setMethod("descriptions", "MultitableResult", function (x) {
+    unlist(lapply(x, function (cube) descriptions(cube)[2]), use.names=FALSE)
 })
 
 #' @rdname show-crunch